Move *Calendar and *Day components to features/
