+++ src/components/features/IndicatorDay.js
